+++ src/main/java/insane96mcp/runeenchanting/RuneHelper.java
@@ -70,8 +70,6 @@ public class RuneHelper {
         stack.set(REDataComponents.RUNES, runes);
         if (!Rune.isCurse(rune))
             stack.set(DataComponents.ENCHANTMENT_GLINT_OVERRIDE, true);
-        if (Rune.isCurse(rune) && getSockets(stack) > 0)
-            stack.set(REDataComponents.SOCKETS, getSockets(stack) + 1);
         return true;
     }
 
@@ -85,17 +83,12 @@ public class RuneHelper {
         }
         else
             stack.set(REDataComponents.RUNES, runes);
-        if (Rune.isCurse(rune) && getSockets(stack) > 0)
-            stack.set(REDataComponents.SOCKETS, getSockets(stack) - 1);
         return true;
     }
 
     public static List<Holder<Rune>> clearRunes(ItemStack stack, boolean clearCurses) {
         List<Holder<Rune>> runes = new ArrayList<>(stack.getOrDefault(REDataComponents.RUNES, List.of()));
         if (clearCurses) {
-            int curseCount = (int) runes.stream().filter(Rune::isCurse).count();
-            if (curseCount > 0 && getSockets(stack) > 0)
-                stack.set(REDataComponents.SOCKETS, Math.max(0, getSockets(stack) - curseCount));
             stack.remove(REDataComponents.RUNES);
         }
         else {
@@ -149,7 +142,7 @@ public class RuneHelper {
     }
 
     public static int getSockets(ItemStack stack) {
-        return stack.getOrDefault(REDataComponents.SOCKETS, 0);
+        return stack.getOrDefault(REDataComponents.SOCKETS, 0) + countCurses(stack);
     }
 
     //TODO Just return an empty list if no component is present